[feat]:다이어리 폰트 파일 추가
