+++ Todoary-iOS/Todoary/CommonSource/UIFont/DiaryFont.swift
@@ -1,0 +1,22 @@
+//
+//  DiaryFont.swift
+//  Todoary
+//
+//  Created by 박지윤 on 2022/08/12.
+//
+
+import Foundation
+
+/*
+ GangwonEduAllBold
+ GangwonEduAllLight
+ 
+ IBMPlexSans-Text
+ IBMPlexSans-Medium
+ 
+ IMHyemin-Regular
+ IMHyemin-Bold
+ 
+ NotoSerifKR-Regular
+ NotoSerifKR-SemiBold
+ */
